Change keys for Presenter and Overview
The o key and p key currently drive the overview and presenter views.
This poses a problem when building an interactive component with a
textbox or textarea within a slide. If the o or p key is pressed while
in a text input, the slideshow will go into that mode, which may be
unexpected behavior.

Require the alt key to be pressed in conjunction with the p key and the
o to trigger the overview and presenter views.