Expose LiveProvider's transformCode prop
`LiveProvider` uses Buble to transform code, but that only supports ES2015, and not newer syntax like Class Properties.  

`LiveProvider` accepts a `transformCode` prop that can be used as an alternative to the built-in transformation, but Spectacle's `ComponentPlayground` doesn't expose that.  By passing `transformCode` through, end users can customize the compilation process (such as using `babel-standalone` instead of Buble).